refactored updateCommands
