+++ src/main/java/seedu/duke/commands/UpdateCommands/UpdateTaskCommand.java
@@ -1,6 +1,8 @@
-package seedu.duke.commands;
+package seedu.duke.commands.UpdateCommands;
 
 import seedu.duke.Ui;
+import seedu.duke.commands.Command;
+import seedu.duke.commands.CommandResult;
 import seedu.duke.exceptions.DukeException;
 import seedu.duke.exceptions.parserexceptions.InvalidBudgetException;
 import seedu.duke.items.Task;
