+++ src/main/java/seedu/duke/parser/commandparser/UpdateParser.java
@@ -3,9 +3,9 @@ package seedu.duke.parser.commandparser;
 import seedu.duke.Duke;
 import seedu.duke.Ui;
 import seedu.duke.commands.Command;
-import seedu.duke.commands.UpdateEventCommand;
-import seedu.duke.commands.UpdateMemberCommand;
-import seedu.duke.commands.UpdateTaskCommand;
+import seedu.duke.commands.UpdateCommands.UpdateEventCommand;
+import seedu.duke.commands.UpdateCommands.UpdateMemberCommand;
+import seedu.duke.commands.UpdateCommands.UpdateTaskCommand;
 import seedu.duke.exceptions.DukeException;
 import seedu.duke.exceptions.parserexceptions.InvalidBudgetException;
 import seedu.duke.items.Event;
