--- src/test/java/seedu/duke/commands/UpdateCommandTest.java
@@ -1,78 +1,0 @@
-package seedu.duke.commands;
-
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-import seedu.duke.exceptions.DukeException;
-import seedu.duke.exceptions.parserexceptions.InvalidBudgetException;
-import seedu.duke.exceptions.parserexceptions.NoCommandAttributesException;
-import seedu.duke.items.Event;
-import seedu.duke.items.Task;
-import seedu.duke.parser.Parser;
-
-import java.io.ByteArrayInputStream;
-import java.io.ByteArrayOutputStream;
-import java.io.InputStream;
-import java.io.PrintStream;
-import java.time.LocalDateTime;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static seedu.duke.Duke.eventCatalog;
-
-public class UpdateCommandTest {
-    private final ByteArrayOutputStream outContent = new ByteArrayOutputStream();
-    private final PrintStream originalOut = System.out;
-    private static ByteArrayInputStream in;
-
-    public void setInput(String input) {
-        in = new ByteArrayInputStream(input.getBytes());
-        System.setIn(in);
-    }
-
-    @BeforeEach
-    public void setUpStream() throws DukeException {
-        System.setOut(new PrintStream(outContent));
-        setUp();
-    }
-
-    @AfterEach
-    public void restoreStream() {
-        System.setOut(originalOut);
-        eventCatalog.clear();
-    }
-
-    @Test
-    //    public void updateCommand_updateNameDateOfEvent_success() throws DukeException, NoCommandAttributesException,
-    //            InvalidBudgetException {
-    //        final InputStream sysInBackup = System.in;
-    //        setInput("title/Charlie Puth Concert> date/21-02-2022 2000");
-    //
-    //        Command command1 = Parser.parseCommand("update 2");
-    //        command1.execute();
-    //        assertEquals("Charlie Puth Concert", eventCatalog.get(1).getTitle());
-    //        assertEquals("21 Feb 2022 - 20:00", eventCatalog.get(1).getStringDateTime());
-    //        System.setIn(sysInBackup);
-    //    }
-
-    void setUp() throws DukeException {
-        LocalDateTime event1DateTime = Parser.convertDateTime("19-02-2022 2000");
-        Event event1 = new Event("Peppa Pig's Concert",
-                "Asia world tour", event1DateTime,
-                "Indoor Stadium", 1000.90);
-        event1.markAsDone();
-        eventCatalog.add(event1);
-        LocalDateTime taskDeadline1 = Parser.convertDateTime("19-02-2022 1930");
-        Task task1 = new Task("Buy Boost", "Need the boost for the concert", taskDeadline1);
-        LocalDateTime taskDeadline2 = Parser.convertDateTime("18-02-2022 1930");
-        Task task2 = new Task("Collect Tickets", "Collection point: Scape", taskDeadline2);
-        LocalDateTime event2DateTime = Parser.convertDateTime("20-02-2022 2030");
-        Event event2 = new Event("Funfair",
-                "For charity", event2DateTime,
-                "Parade square", 2000.10);
-        eventCatalog.add(event2);
-        eventCatalog.get(0).addToTaskList(task1);
-        eventCatalog.get(0).addToTaskList(task2);
-        eventCatalog.sortCatalog();
-    }
-}
